# runtime/examples/ReactFastApi.jsplugin/backend/openapi.json
{
  "components": {
    "schemas": {
      "ActorResponse": {
        "properties": {
          "permissions": {
            "items": {
              "enum": [
                "profile.read",
                "shop.purchase",
                "realtime.connect",
                "permissions.read",
                "permissions.manage"
              ],
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "member",
                "moderator",
                "admin"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "player_id",
          "player_name",
          "roles",
          "permissions"
        ],
        "title": "ActorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionRequest": {
        "properties": {
          "player_id": {
            "maxLength": 36,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9:_-]{0,35}$",
            "title": "Player Id",
            "type": "string"
          },
          "player_name": {
            "maxLength": 16,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_]{1,16}$",
            "title": "Player Name",
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "player_name"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/ActorResponse"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "actor"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Graaly game API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/me": {
      "get": {
        "operationId": "current_actor_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve the authenticated player's current roles and permissions",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/session": {
      "post": {
        "operationId": "create_session_v1_auth_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-graaly-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Graaly-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange a trusted game-server identity for a short player session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}